backend/requirements.txt
Django==3.2.2
psycopg2-binary>=2.8
djangorestframework==3.12.4
mysqlclient==2.0.3
django-mysql==3.11.1
django-cors-headers==3.7.0
Pillow==8.2.0
PyJWT==2.1.0
Faker==8.1.2
django-redis==4.12.1
celery==4.4.2
